23 04 03 정승원 기술스택 vo languageMapper

+++ alli/src/main/java/com/smhrd/model/FrameworkVO.java
@@ -1,8 +1,10 @@
 package com.smhrd.model;
 
+import java.math.BigDecimal;
+
 public class FrameworkVO {
 	
-	private int resume_num;
+	private BigDecimal resume_num;
 	private String framework_name;
 	private String framework_level;
 	
@@ -10,14 +12,14 @@ public class FrameworkVO {
 		
 	}
 
-	public FrameworkVO(int resume_num, String framework_name, String framework_level) {
+	public FrameworkVO(BigDecimal resume_num, String framework_name, String framework_level) {
 		super();
 		this.resume_num = resume_num;
 		this.framework_name = framework_name;
 		this.framework_level = framework_level;
 	}
 
-	public int getResume_num() {
+	public BigDecimal getResume_num() {
 		return resume_num;
 	}
